+++ src/domain/logs/BaseLogger.ts
